requirements.txt
anyio>=4.8.0,<4.9.0
Jinja2
SQLAlchemy>=2.0.0,<2.1.0
cobra>=0.14.2,<0.30.0
python-libsbml>=5.20.0,<5.21.0
numpy>=2.2.0,<2.3.0
pandas>=2.2.0,<2.3.0
psycopg2-binary>=2.9.0,<2.10.0
biopython>=1.85,<1.86
scipy>=1.15.0,<1.16.0
lxml>=5.3.0,<5.4.0
pytest>=8.3.0,<8.4.0
escher>=1.8.0,<1.9.0
configparser>=7.1.0,<7.2.0
 
